Single-stepping with QEMU and GDB
It's possible to put the ARM emulator qemu-arm into a mode where it
accepts commands from an instance of the Gnu debugger gdb running on
the host machine. This makes it possible to run an ARM program
written in assembly language and trace its action one instruction at
a time.